backend/requirements.txt
appnope==0.1.4
asttokens==3.0.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.12
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
Flask==3.1.0
flask-cors==5.0.1
idna==3.10
ipykernel==6.29.5
ipython==9.0.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kagglehub==0.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
six==1.17.0
spotipy==2.25.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
